requirements.txt
absl-py==1.4.0
altgraph==0.17.3
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
customtkinter==5.2.0
cycler==0.11.0
darkdetect==0.8.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.7
docopt==0.6.2
executing==1.2.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.41.1
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
idna==3.4
ipykernel==6.25.0
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.13.1
keyboard==0.13.5
kiwisolver==1.4.4
labelImg==1.8.6
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mediapipe==0.10.3
mpmath==1.3.0
mysql-connector-python==8.0.33
nest-asyncio==1.5.7
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
openvino==2023.0.1
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==10.0.0
pipreqs==0.4.13
platformdirs==3.9.1
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
sounddevice==0.4.6
stack-data==0.6.2
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-hub==0.14.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thop==0.1.1.post2209072238
tk==0.1.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.143
urllib3==1.26.16
virtualenv==20.24.3
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.15.0
yarg==0.1.9
